✨ feat: Gates e Polices on super-admin RegisterPermission actions

+++ app/Http/Controllers/Api/UserController.php
@@ -10,9 +10,12 @@ use App\Library\Converters\Phone as PhoneConverter;
 use App\Models\RegisterPermission;
 use App\Models\User;
 use Illuminate\Auth\Events\Registered;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
+
     public function __construct(private readonly RegisterServiceInterface $registerService)
     {
         // ...
@@ -26,8 +29,11 @@ class UserController extends Controller
         $registerPermission = RegisterPermission::where('email', $request->email)->first();
         $phone = PhoneConverter::clear($registerPermission->phone ?? $request->phone);
         $fields = [...$request->only(['name', 'email', 'password']), 'phone' => $phone];
-        $user = User::create($fields);
+
+        $this->authorize('delete', $registerPermission);
         RegisterPermission::destroy($registerPermission->id);
+
+        $user = User::create($fields);
         event(new Registered($user));
         // $this->applyDefaultUserRole($user);
